command-service/requirements.txt
# -----------------------------------------------------------------------------
# ARQUIVO: requirements.txt
# -----------------------------------------------------------------------------
fastapi
uvicorn[standard]
requests
paho-mqtt
